Move Signature in its own module
